Always allocate cells for Futhark (cuda backend) temporaries from the front of the GPU mem pool.
The old behavior put temporaries behind the back cursor of the
GPU mem pool, while selectively inferring which cells to push
to the front by introspecting on Futhark memblock tags. While
this helped to minimize mem pool fragmentation, subsequent
investigation has determined the front-back partition of the
mem pool to be detrimental to serious usage of GPU soft-OOM.
So, simply do not allocate from the back cursor in the cuda
backend thunk impl.

As a bonus, this also fixes a thrashing-related performance
degradation observed when interleaving OOM gradient updates
and single forward passes.